requirements.txt
alabaster==0.7.11
argon2-cffi==18.1.0
astroid==2.0.4
Babel==2.6.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.8
django-bootstrap4==0.0.6
django-haystack==2.8.1
djangorestframework==3.8.2
docutils==0.14
idna==2.7
imagesize==1.0.0
isodate==0.6.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.0
mccabe==0.6.1
mongoengine==0.15.3
mysqlclient==1.3.13
networkx==2.2
packaging==17.1
prov==1.5.2
pycparser==2.18
Pygments==2.2.0
pylint==2.1.1
pylint-django==2.0
pylint-plugin-utils==0.4
pymongo==3.7.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
rdflib==4.2.2
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
typed-ast==1.1.0
urllib3==1.23
whoosh==2.7.4
wrapt==1.10.11
